CATTY-684 Add body pose detection support for older iOS versions
