+++ src/Catty/Settings.bundle/Acknowledgements.plist
@@ -112,6 +112,12 @@
                 <key>Type</key>
                 <string>PSGroupSpecifier</string>
                 <key>FooterText</key>
+                <string>Catty/Resources/Licenses/PoseNet</string>
+            </dict>
+            <dict>
+                <key>Type</key>
+                <string>PSGroupSpecifier</string>
+                <key>FooterText</key>
                 <string>Catty/Resources/Licenses/SPUserResizableView</string>
             </dict>
             <dict>
