+++ src/Catty/Setup/CatrobatSetup+Sensors.swift
@@ -126,19 +126,21 @@ extension CatrobatSetup {
         if #available(iOS 14.0, *) {
             sensorArray.append(NeckXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
             sensorArray.append(NeckYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftShoulderXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftShoulderYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightShoulderXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightShoulderYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftElbowXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftElbowYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightElbowXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightElbowYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftWristXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftWristYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightWristXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightWristYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        }
+        sensorArray.append(LeftShoulderXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftShoulderYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightShoulderXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightShoulderYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftElbowXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftElbowYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightElbowXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightElbowYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftWristXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftWristYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightWristXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightWristYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
 
+        if #available(iOS 14.0, *) {
             sensorArray.append(LeftPinkyKnuckleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
             sensorArray.append(LeftPinkyKnuckleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
             sensorArray.append(RightPinkyKnuckleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
@@ -159,20 +161,20 @@ extension CatrobatSetup {
             sensorArray.append(LeftThumbKnuckleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
             sensorArray.append(RightThumbKnuckleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
             sensorArray.append(RightThumbKnuckleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-
-            sensorArray.append(LeftHipXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftHipYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightHipXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightHipYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftKneeXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftKneeYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightKneeXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightKneeYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftAnkleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(LeftAnkleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightAnkleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
-            sensorArray.append(RightAnkleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
         }
+
+        sensorArray.append(LeftHipXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftHipYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightHipXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightHipYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftKneeXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftKneeYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightKneeXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightKneeYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftAnkleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(LeftAnkleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightAnkleXSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
+        sensorArray.append(RightAnkleYSensor(stageSize: stageSize, visualDetectionManagerGetter: { visualDetectionManager }))
 
         if #available(iOS 13.0, *) {
             sensorArray.append(TextFromCameraSensor(visualDetectionManagerGetter: { visualDetectionManager }))
